requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alluka==0.3.3
attrs==25.4.0
cffi==2.0.0
colorlog==6.10.1
frozenlist==1.8.0
hikari==2.5.0
hikari-arc==2.1.1
idna==3.11
iso8601==2.1.0
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
pynacl==1.6.1
pypika-tortoise==0.6.2
pytz==2025.2
tortoise-orm==0.25.1
typing-extensions==4.15.0
uvloop==0.22.1
yarl==1.22.0
